feat(dashboard/ui): implement initial dashboard and settings pages with their UI components
