+++ apps/client/src/features/settings/pages/index.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import { useState } from 'react';
 import { Camera } from 'lucide-react';
 import { ProfileForm } from '../components/ProfileForm';
 import { APIKeysConfig } from '../components/APIKeysConfig';
